backend/requirements.txt
# Django & REST Framework
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# CORS
django-cors-headers==4.3.1

# Environment variables
python-dotenv==1.0.0

# Password validation
argon2-cffi==23.1.0

# Email
django-ses==3.5.2

# Utilities
pytz==2024.1
python-dateutil==2.8.2

# CSV handling
chardet==5.2.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Development tools
ipython==8.20.0
django-extensions==3.2.3
